HBASE-1262 Eclipse warnings, including performance related things like synthetic accessors
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/trunk@755878 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/hbase/metrics/MetricsRate.java
@@ -53,7 +53,7 @@ public class MetricsRate {
     long now = System.currentTimeMillis();
     long diff = (now-ts)/1000;
     if (diff == 0) diff = 1; // sigh this is crap.
-    this.prevRate = value / diff;
+    this.prevRate = (float)value / diff;
     this.value = 0;
     this.ts = now;
   }
